fix: position scroll-to-bottom button above chat input bar
