+++ src/components/ai-chat/scroll-to-bottom-button.tsx
@@ -37,8 +37,8 @@ export function ScrollToBottomButton({
 
 const styles = stylex.create({
   button: {
-    position: "fixed",
-    bottom: `calc(${space._10} + env(safe-area-inset-bottom))`,
+    position: "absolute",
+    bottom: `calc(100% + ${space._2})`,
     left: "50%",
     transform: "translateX(-50%)",
     display: "inline-flex",
